Fix issue #4, jump to last line
This commit addresses an issue in caret positioning where caret would shift even in the last line of the document if it was in the upper half - which isn’t ideal behavior. We now check if it's the last line and in the upper half before deciding whether a shift is required or not. This fixes issue #4 raised on GitHub.