Refactor caret scroll logic for customizable padding unit
Incorporated the choice of padding unit in `ScrollCaretListener` to deliver a more adaptable user experience. Padding unit, which can be either in 'Lines' or 'Relative(%)', is now a selectable option due to the introduction of `PaddingUnit` enum in `ScrollPluginSettings`. The scrolling logic has been refactored accordingly to handle the chosen padding unit, and if both 'top' and 'bottom' settings lead to collision at the center, scrolling will default to center. User input for relative padding units is also handled within the UI through changes in `ScrollPluginConfigurable` and `PluginSettingsComponent`.